# api/openapi.yaml
openapi: 3.0.3
info:
  title: IGNAI Hackathon Platform API
  version: 0.4.0
  description: 'Spring Boot / MyBatis 版本接口规划。MVP 无独立评委角色，无视频上传； 后台管理入口规划为 /root/，root
    初始化和认证使用 /api/root/*。

    '
servers:
- url: http://localhost:8080/api
  description: Local Spring Boot API
...
paths:
...
  /admin/dashboard:
    get:
      tags:
      - Admin
      summary: 后台概览统计
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 统计数据
          content:
            application/json:
              schema:
                type: object
      operationId: admin_dashboard_get
  /admin/users:
    get:
      tags:
      - Admin
      summary: 用户表
      security:
      - bearerAuth: []
      parameters:
      - name: role
        in: query
        schema:
          type: string
          enum:
          - root
          - admin
          - participant
      - name: status
        in: query
        schema:
          type: string
          enum:
          - active
          - disabled
      - name: keyword
        in: query
        schema:
          type: string
      responses:
        '200':
          description: 用户列表
          content:
            application/json:
              schema:
                type: object
      operationId: admin_users_get
  /admin/users/{userId}:
    get:
      tags:
      - Admin
      summary: 用户详情
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: 用户详情
          content:
            application/json:
              schema:
                type: object
      operationId: admin_users_userId_get
  /admin/users/{userId}/status:
    patch:
      tags:
      - Admin
      summary: 启用或禁用用户
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - status
              properties:
                status:
                  type: string
                  enum:
                  - active
                  - disabled
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
      operationId: admin_users_userId_status_patch
  /admin/users/admins:
    post:
      tags:
      - Admin
      summary: root 创建管理员账号
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
      operationId: admin_users_admins_post
  /admin/events:
    get:
      tags:
      - Admin
      summary: 活动列表
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 活动列表
          content:
            application/json:
              schema:
                type: object
      operationId: admin_events_get
    post:
      tags:
      - Admin
      summary: 创建活动
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventUpsertRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
      operationId: admin_events_post
  /admin/events/{eventId}:
    put:
      tags:
      - Admin
      summary: 更新活动
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventUpsertRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
      operationId: admin_events_eventId_put
  /admin/events/{eventId}/windows:
    patch:
      tags:
      - Admin
      summary: 设置报名、提交、评分和点赞窗口
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventWindowRequest'
      responses:
        '200':
          description: 保存成功
          content:
            application/json:
              schema:
                type: object
      operationId: admin_events_eventId_windows_patch
  /admin/events/{eventId}/tracks:
    get:
      tags:
      - Admin
      summary: 赛道列表
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: 赛道列表
          content:
            application/json:
              schema:
                type: object
      operationId: admin_events_eventId_tracks_get
    post:
      tags:
      - Admin
      summary: 新增赛道
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackCreateRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
      operationId: admin_events_eventId_tracks_post
...
  /admin/registrations:
    get:
      tags:
      - Registration
      summary: 报名列表
      security:
      - bearerAuth: []
      parameters:
      - name: status
        in: query
        schema:
          type: string
          enum:
          - pending
          - approved
          - rejected
          - waitlist
          - cancelled
      - name: trackId
        in: query
        schema:
          type: integer
      - name: aiTag
        in: query
        schema:
          type: string
      - name: keyword
        in: query
        schema:
          type: string
      responses:
        '200':
          description: 报名列表
          content:
            application/json:
              schema:
                type: object
      operationId: admin_registrations_get
  /admin/registrations/{registrationId}:
    get:
      tags:
      - Registration
      summary: 报名详情
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/RegistrationId'
      responses:
        '200':
          description: 报名详情
          content:
            application/json:
              schema:
                type: object
      operationId: admin_registrations_registrationId_get
...
  /admin/registrations/{registrationId}/review:
    patch:
      tags:
      - Registration
      summary: 审核报名
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/RegistrationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationReviewRequest'
      responses:
        '200':
          description: 审核成功
          content:
            application/json:
              schema:
                type: object
      operationId: admin_registrations_registrationId_review_patch
  /admin/registrations/batch-review:
    post:
      tags:
      - Registration
      summary: 批量审核报名
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - registrationIds
              - status
              properties:
                registrationIds:
                  type: array
                  items:
                    type: integer
                status:
                  type: string
                  enum:
                  - approved
                  - rejected
                  - waitlist
                note:
                  type: string
      responses:
        '200':
          description: 批量审核成功
          content:
            application/json:
              schema:
                type: object
      operationId: admin_registrations_batch-review_post
...
  /projects:
    post:
      tags:
      - Project
      summary: 创建或保存作品
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpsertRequest'
      responses:
        '200':
          description: 作品信息
          content:
            application/json:
              schema:
                type: object
      operationId: projects_post
...
  /projects/{projectId}:
    put:
      tags:
      - Project
      summary: 更新作品
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
      operationId: projects_projectId_put
...
  /admin/projects:
    get:
      tags:
      - Project
      summary: 后台项目看板
      security:
      - bearerAuth: []
      parameters:
      - name: status
        in: query
        schema:
          type: string
          enum:
          - draft
          - submitted
          - locked
          - published
          - archived
      - name: trackId
        in: query
        schema:
          type: integer
      - name: sort
        in: query
        schema:
          type: string
          enum:
          - latest
          - votes
          - rating
      - name: keyword
        in: query
        schema:
          type: string
      responses:
        '200':
          description: 项目列表
          content:
            application/json:
              schema:
                type: object
      operationId: admin_projects_get
  /admin/projects/{projectId}:
    get:
      tags:
      - Project
      summary: 后台项目详情
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 项目详情
          content:
            application/json:
              schema:
                type: object
      operationId: admin_projects_projectId_get
  /admin/projects/{projectId}/status:
    patch:
      tags:
      - Project
      summary: 更新项目状态
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - status
              properties:
                status:
                  type: string
                  enum:
                  - draft
                  - submitted
                  - locked
                  - published
                  - archived
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
      operationId: admin_projects_projectId_status_patch
  /admin/projects/{projectId}/showcase:
    patch:
      tags:
      - Project
      summary: 设置作品展示规则
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectShowcaseRequest'
      responses:
        '200':
          description: 保存成功
          content:
            application/json:
              schema:
                type: object
      operationId: admin_projects_projectId_showcase_patch
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        type: integer
    RegistrationId:
      name: registrationId
      in: path
      required: true
      schema:
        type: integer
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    AdminCreateRequest:
      type: object
      required:
      - name
      - email
      - password
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 8
...
    EventUpsertRequest:
      type: object
      required:
      - title
      properties:
        title:
          type: string
        subtitle:
          type: string
        location:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
          - draft
          - published
          - running
          - archived
        officialSiteUrl:
          type: string
        benchmarkSiteUrl:
          type: string
    EventWindowRequest:
      type: object
      properties:
        registrationOpen:
          type: boolean
        ratingEnabled:
          type: boolean
        voteEnabled:
          type: boolean
        registrationDeadline:
          type: string
          format: date-time
        submissionDeadline:
          type: string
          format: date-time
        ratingStartAt:
          type: string
          format: date-time
        ratingEndAt:
          type: string
          format: date-time
    TrackCreateRequest:
      type: object
      required:
      - name
      properties:
        name:
          type: string
        description:
          type: string
        sortOrder:
          type: integer
...
    RegistrationReviewRequest:
      type: object
      required:
      - status
      properties:
        status:
          type: string
          enum:
          - approved
          - rejected
          - waitlist
        note:
          type: string
        generateMailDraft:
          type: boolean
...
    ProjectUpsertRequest:
      type: object
      required:
      - eventId
      - teamId
      - title
      properties:
        eventId:
          type: integer
        teamId:
          type: integer
        trackId:
          type: integer
        title:
          type: string
        tagline:
          type: string
        description:
          type: string
        demoUrl:
          type: string
        codeUrl:
          type: string
        payload:
          type: object
    ProjectShowcaseRequest:
      type: object
      properties:
        published:
          type: boolean
        featured:
          type: boolean
        sortWeight:
          type: integer
        showcaseNote:
          type: string
...